requirements.txt
-i https://pypi.org/simple

certifi==2020.12.5

chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

flask-login==0.5.0

flask-sqlalchemy==2.5.1

flask-wtf==0.14.3

flask==1.1.2

greenlet==1.0.0; python_version >= '3'

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

jinja2==2.11.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

passlib==1.7.4

psycopg2-binary==2.8.6

psycopg2==2.8.6

python-dotenv==0.16.0

requests==2.25.1

sqlalchemy==1.4.4

urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

wtforms==2.3.3

gunicorn==19.3.0
